+++ cpp/cluster-controller/http-communication-manager/LongPollingMessageReceiver.cpp
@@ -28,9 +28,7 @@
 #include "joynr/DispatcherUtils.h"
 #include "joynr/TypeUtil.h"
 #include "cluster-controller/httpnetworking/HttpResult.h"
-#include "joynr/ILocalChannelUrlDirectory.h"
 #include "joynr/Future.h"
-#include "joynr/types/ChannelUrlInformation.h"
 #include "joynr/JoynrMessage.h"
 #include "joynr/system/RoutingTypes/ChannelAddress.h"
 #include "joynr/MessageRouter.h"
@@ -47,7 +45,6 @@ LongPollingMessageReceiver::LongPollingMessageReceiver(
         const std::string& receiverId,
         const LongPollingMessageReceiverSettings& settings,
         Semaphore* channelCreatedSemaphore,
-        std::shared_ptr<ILocalChannelUrlDirectory> channelUrlDirectory,
         std::function<void(const std::string&)> onTextMessageReceived)
         : Thread("LongPollRecv"),
           brokerUrl(brokerUrl),
@@ -57,7 +54,6 @@ LongPollingMessageReceiver::LongPollingMessageReceiver(
           interrupted(false),
           interruptedMutex(),
           interruptedWait(),
-          channelUrlDirectory(channelUrlDirectory),
           channelCreatedSemaphore(channelCreatedSemaphore),
           onTextMessageReceived(onTextMessageReceived)
 {
@@ -109,18 +105,8 @@ void LongPollingMessageReceiver::run()
             interruptedWait.wait_for(lock, settings.createChannelRetryInterval);
         }
     }
-    /**
-      * register the channelUrl with the ChannelUrlDirectory (asynchronously)
-      */
-    assert(channelUrlDirectory != nullptr);
-    types::ChannelUrlInformation urlInformation;
-    std::vector<std::string> urls = {channelUrl};
-    urlInformation.setUrls(urls);
-    JOYNR_LOG_INFO(
-            logger,
-            "Adding channelId and Url of cluster controller to remote ChannelUrlDirectory {}",
-            channelUrl);
-    channelUrlDirectory->registerChannelUrlsAsync(channelId, urlInformation);
+
+    // TODO: The received URL must be forwarded in such a way that ChannelAddress objects use it.
 
     while (!isInterrupted()) {