+++ cpp/cluster-controller/include/joynr/IMessageReceiver.h
@@ -26,8 +26,6 @@
 namespace joynr
 {
 
-class ILocalChannelUrlDirectory;
-
 /**
   * @class IMessageReceiver
   * @brief Interface for message receiver
@@ -73,8 +71,6 @@ public:
       * Will try to delete the channel from the server. Returns true if successfull, false if not.
       */
     virtual bool tryToDeleteChannel() = 0;
-
-    virtual void init(std::shared_ptr<ILocalChannelUrlDirectory> channelUrlDirectory) = 0;
 
     virtual void registerReceiveCallback(
             std::function<void(const std::string&)> onTextMessageReceived) = 0;