+++ cpp/libjoynr/joynr-messaging/MessagingSettings.cpp
@@ -57,24 +57,6 @@ const std::string& MessagingSettings::SETTING_DISCOVERY_DIRECTORIES_DOMAIN()
     return value;
 }
 
-const std::string& MessagingSettings::SETTING_CHANNEL_URL_DIRECTORY_URL()
-{
-    static const std::string value("messaging/channel-url-directory-url");
-    return value;
-}
-
-const std::string& MessagingSettings::SETTING_CHANNEL_URL_DIRECTORY_CHANNELID()
-{
-    static const std::string value("messaging/channel-url-directory-channelid");
-    return value;
-}
-
-const std::string& MessagingSettings::SETTING_CHANNEL_URL_DIRECTORY_PARTICIPANTID()
-{
-    static const std::string value("messaging/channel-url-directory-participantid");
-    return value;
-}
-
 const std::string& MessagingSettings::SETTING_CAPABILITIES_DIRECTORY_URL()
 {
     static const std::string value("messaging/capabilities-directory-url");
@@ -322,21 +304,6 @@ std::string MessagingSettings::getDiscoveryDirectoriesDomain() const
     return settings.get<std::string>(SETTING_DISCOVERY_DIRECTORIES_DOMAIN());
 }
 
-std::string MessagingSettings::getChannelUrlDirectoryUrl() const
-{
-    return settings.get<std::string>(SETTING_CHANNEL_URL_DIRECTORY_URL());
-}
-
-std::string MessagingSettings::getChannelUrlDirectoryChannelId() const
-{
-    return settings.get<std::string>(SETTING_CHANNEL_URL_DIRECTORY_CHANNELID());
-}
-
-std::string MessagingSettings::getChannelUrlDirectoryParticipantId() const
-{
-    return settings.get<std::string>(SETTING_CHANNEL_URL_DIRECTORY_PARTICIPANTID());
-}
-
 std::string MessagingSettings::getCapabilitiesDirectoryUrl() const
 {
     return settings.get<std::string>(SETTING_CAPABILITIES_DIRECTORY_URL());
@@ -539,16 +506,6 @@ void MessagingSettings::checkSettings()
 
     assert(settings.contains(SETTING_DISCOVERY_DIRECTORIES_DOMAIN()));
 
-    assert(settings.contains(SETTING_CHANNEL_URL_DIRECTORY_URL()));
-    std::string channelUrlDirectoryUrl =
-            settings.get<std::string>(SETTING_CHANNEL_URL_DIRECTORY_URL());
-    if (channelUrlDirectoryUrl.back() != '/') {
-        channelUrlDirectoryUrl.append("/");
-        settings.set(SETTING_CHANNEL_URL_DIRECTORY_URL(), channelUrlDirectoryUrl);
-    }
-    assert(settings.contains(SETTING_CHANNEL_URL_DIRECTORY_CHANNELID()));
-    assert(settings.contains(SETTING_CHANNEL_URL_DIRECTORY_PARTICIPANTID()));
-
     assert(settings.contains(SETTING_CAPABILITIES_DIRECTORY_URL()));
     std::string capabilitiesDirectoryUrl =
             settings.get<std::string>(SETTING_CAPABILITIES_DIRECTORY_URL());
@@ -599,18 +556,6 @@ void MessagingSettings::printSettings() const
                     "SETTING: {} = {})",
                     SETTING_DISCOVERY_DIRECTORIES_DOMAIN(),
                     settings.get<std::string>(SETTING_DISCOVERY_DIRECTORIES_DOMAIN()));
-    JOYNR_LOG_DEBUG(logger,
-                    "SETTING: {}  = {})",
-                    SETTING_CHANNEL_URL_DIRECTORY_URL(),
-                    settings.get<std::string>(SETTING_CHANNEL_URL_DIRECTORY_URL()));
-    JOYNR_LOG_DEBUG(logger,
-                    "SETTING: {}  = {})",
-                    SETTING_CHANNEL_URL_DIRECTORY_CHANNELID(),
-                    settings.get<std::string>(SETTING_CHANNEL_URL_DIRECTORY_CHANNELID()));
-    JOYNR_LOG_DEBUG(logger,
-                    "SETTING: {}  = {})",
-                    SETTING_CHANNEL_URL_DIRECTORY_PARTICIPANTID(),
-                    settings.get<std::string>(SETTING_CHANNEL_URL_DIRECTORY_PARTICIPANTID()));
     JOYNR_LOG_DEBUG(logger,
                     "SETTING: {}  = {})",
                     SETTING_CAPABILITIES_DIRECTORY_URL(),