+++ cpp/tests/utils/MockObjects.h
@@ -75,19 +75,16 @@
 #include "joynr/InProcessConnectorFactory.h"
 #include "cluster-controller/http-communication-manager/HttpReceiver.h"
 
-#include "joynr/infrastructure/ChannelUrlDirectoryProxy.h"
 #include "joynr/infrastructure/GlobalDomainAccessControllerProxy.h"
 
 #include "joynr/MessageRouter.h"
 
-#include "joynr/ILocalChannelUrlDirectory.h"
 #include "joynr/LocalCapabilitiesDirectory.h"
 #include "joynr/ParticipantIdStorage.h"
 #include "joynr/MessagingSettings.h"
 #include "joynr/SubscriptionManager.h"
 #include "joynr/PublicationManager.h"
 #include "joynr/DiscoveryQos.h"
-#include "joynr/types/ChannelUrlInformation.h"
 #include "joynr/IMessageSender.h"
 #include "joynr/BrokerUrl.h"
 #include "joynr/Directory.h"
@@ -526,7 +523,7 @@ class MockMessageReceiver : public joynr::IMessageReceiver
 {
 public:
     MockMessageReceiver() = default;
-    MOCK_METHOD1(init, void(std::shared_ptr<joynr::ILocalChannelUrlDirectory> channelUrlDirectory));
+    MOCK_METHOD0(init, void());
     MOCK_CONST_METHOD0(getGlobalClusterControllerAddress, std::string&());
     MOCK_METHOD0(startReceiveQueue, void());
     MOCK_METHOD0(stopReceiveQueue, void());
@@ -540,7 +537,7 @@ class MockMessageSender : public joynr::IMessageSender
 {
 public:
     MOCK_METHOD3(sendMessage,void(const joynr::system::RoutingTypes::Address&, const joynr::JoynrMessage&, const std::function<void(const joynr::exceptions::JoynrRuntimeException&)>&));
-    MOCK_METHOD2(init,void(std::shared_ptr<joynr::ILocalChannelUrlDirectory> channelUrlDirectory,const joynr::MessagingSettings& settings));
+    MOCK_METHOD1(init,void(const joynr::MessagingSettings& settings));
     MOCK_METHOD1(registerReceiveQueueStartedCallback, void(std::function<void(void)> waitForReceiveQueueStarted));
 };
 
@@ -784,71 +781,6 @@ public:
 class MockPublicationManager : public joynr::PublicationManager {
 public:
     MOCK_METHOD2(attributeValueChanged, void(const std::string& subscriptionId, const joynr::Variant& value));
-};
-
-class MockChannelUrlDirectoryProxy : public virtual joynr::infrastructure::ChannelUrlDirectoryProxy {
-public:
-    MockChannelUrlDirectoryProxy() :
-        ChannelUrlDirectoryProxy(std::make_shared<const joynr::system::RoutingTypes::Address>(), nullptr, nullptr, "domain", joynr::MessagingQos(), false),
-        ProxyBase(nullptr, nullptr, "domain", joynr::MessagingQos(), false),
-        ChannelUrlDirectoryProxyBase(std::make_shared<const joynr::system::RoutingTypes::Address>(), nullptr, nullptr, "domain", joynr::MessagingQos(), false),
-        ChannelUrlDirectorySyncProxy(std::make_shared<const joynr::system::RoutingTypes::Address>(), nullptr, nullptr, "domain", joynr::MessagingQos(), false),
-        ChannelUrlDirectoryAsyncProxy(std::make_shared<const joynr::system::RoutingTypes::Address>(), nullptr, nullptr, "domain", joynr::MessagingQos(), false){}
-
-    MOCK_METHOD3(getUrlsForChannelAsync,
-                 std::shared_ptr<joynr::Future<joynr::types::ChannelUrlInformation>> (
-                     const std::string& channelId,
-                     std::function<void(const joynr::types::ChannelUrlInformation& urls)> onSuccess,
-                     std::function<void(const joynr::exceptions::JoynrRuntimeException& error)> onRuntimeError
-                 )
-    );
-
-    MOCK_METHOD4(registerChannelUrlsAsync,
-                 std::shared_ptr<joynr::Future<void> >(
-                     const std::string& channelId,
-                     const joynr::types::ChannelUrlInformation& channelUrlInformation,
-                     std::function<void(void)> onSuccess,
-                     std::function<void(const joynr::exceptions::JoynrRuntimeException& error)> onRuntimeError
-                 )
-    );
-
-    MOCK_METHOD3(unregisterChannelUrlsAsync,
-                 std::shared_ptr<joynr::Future<void>>(
-                     const std::string& channelId,
-                     std::function<void(void)> onSuccess,
-                     std::function<void(const joynr::exceptions::JoynrRuntimeException& error)> onRuntimeError
-                 )
-    );
-};
-
-
-class MockLocalChannelUrlDirectory : public joynr::ILocalChannelUrlDirectory {
-public:
-    MOCK_METHOD4(registerChannelUrlsAsync,
-                 std::shared_ptr<joynr::Future<void>>(
-                     const std::string& channelId,
-                     joynr::types::ChannelUrlInformation channelUrlInformation,
-                     std::function<void(void)> onSuccess,
-                     std::function<void(const joynr::exceptions::JoynrRuntimeException& error)> onError
-                 )
-    );
-
-    MOCK_METHOD3(unregisterChannelUrlsAsync,
-                 std::shared_ptr<joynr::Future<void>>(
-                     const std::string& channelId,
-                     std::function<void(void)> onSuccess,
-                     std::function<void(const joynr::exceptions::JoynrRuntimeException& error)> onError
-                 )
-    );
-
-    MOCK_METHOD4(getUrlsForChannelAsync,
-                 std::shared_ptr<joynr::Future<joynr::types::ChannelUrlInformation>>(
-                     const std::string& channelId,
-                     const std::chrono::milliseconds timeout,
-                     std::function<void(const joynr::types::ChannelUrlInformation&)> onSuccess,
-                     std::function<void(const joynr::exceptions::JoynrRuntimeException& error)> onError
-                 )
-    );
 };
 
 class MockParticipantIdStorage : public joynr::ParticipantIdStorage {